+++ cmd/sfreleaser/release_github.go
@@ -6,32 +6,33 @@ import (
 	"github.com/streamingfast/cli"
 )
 
-func releaseGithub(model *GitHubReleaseModel) {
+func releaseGithub(global *GlobalModel, release *ReleaseModel, githubRelease *GitHubReleaseModel) {
 	if devSkipGoreleaser {
 		return
 	}
 
+	renderGoreleaserFile(global, release, githubRelease)
+
 	arguments := []string{
 		"docker",
 
-		// docker arguments
 		"run",
 		"--rm",
 		"-e CGO_ENABLED=1",
-		"--env-file", model.EnvFilePath,
+		"--env-file", githubRelease.EnvFilePath,
 		"-v /var/run/docker.sock:/var/run/docker.sock",
 		"-v", cli.WorkingDirectory() + ":/go/src/work",
 		"-w /go/src/work",
-		model.GoreleaserImageID,
+		githubRelease.GoreleaserImageID,
 
 		// goreleaser arguments
-		"-f", model.GoreleaseConfigPath,
+		"-f", githubRelease.GoreleaseConfigPath,
 		"--timeout=60m",
 		"--rm-dist",
-		"--release-notes=" + model.ReleaseNotesPath,
+		"--release-notes=" + githubRelease.ReleaseNotesPath,
 	}
 
-	if model.AllowDirty {
+	if githubRelease.AllowDirty {
 		arguments = append(arguments, "--skip-validate")
 	}