+++ cmd/sfreleaser/release_golang.go
@@ -1,3 +1,0 @@
-package main
-
-var releaseGolangGitHub = releaseGithub