+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.5.0
+
+* Removed the need to have `.goreleaser.yaml` file in the repository (file is now generated on the fly).
+
+* Added support for disabling Brew tap release (enabled by default).
+
+* Added support for specifying `owner` (defaults to `streamingfast`).
+
+* Added support for specifying `license` (defaults to `Apache-2.0`).
+
 ## v0.4.2
 
 * Added support for resolving files relative to `.sfreleaser` location.